alembic/versions/7ba43896e8f1_init_db.py
"""init db

Revision ID: 7ba43896e8f1
Revises: 
Create Date: 2024-10-15 11:47:20.040552

"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa


# revision identifiers, used by Alembic.
revision: str = '7ba43896e8f1'
down_revision: Union[str, None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    pass


def downgrade() -> None:
    pass
