pyproject.toml
[project]
name = "pouetbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.3",
    "discord-py>=2.4.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
]
